More code for performance and such
This is honestly pretty messy. Hopefully I take the time in the future
to come back and clean this up as well as describe why everything is
here and what I did what I did. But hey, we are at between -46.59% and
-32.12% cycle reduction from normal exceptions which I am quite proud
of. All that is needed is to weaken the symbols used in
except_experimental.cpp in your libgcc.a file, or any others that cause
link time failures, and then use the implementations here and you get
that reduction.